Update patch set 4
Patch Set 4:

(2 comments)

Here’s another thought: is the value of %ebx even relevant to the stackwalk at any point that we’ve seen? If not, then we can just prime the postfix evaluator’s dictionary with d["$ebx"] = 0 if there’s nothing valid to put there so that if something does reference it, evaluation at least won’t fail just for that reason.

We’d accidentally wind up seeing %ebx as valid output of program string evaluation if we ever saw a program string that had “$ebx $ebx =” in it, but if we’re saying that %ebx isn’t really ever relevant to the walk anyway, then perhaps it doesn’t matter.

Patch-set: 4
Label: Code-Review=0