Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (2 comments)
> 
> Here’s another thought: is the value of %ebx even relevant to the stackwalk at any point that we’ve seen? If not, then we can just prime the postfix evaluator’s dictionary with d["$ebx"] = 0 if there’s nothing valid to put there so that if something does reference it, evaluation at least won’t fail just for that reason.

Do you mean when we don't have a (context_validity & EBX), or always?

The correct value of EBX is necessary for the the function I'm looking at (RtlSleepConditionVariableSRW). It uses it in an EBP-alike way for part of the function. Setting it to 0 causes the evaluator to fail (causing the stack walk to fail) when it tries to deference ebx+4 to get stack data.

> 
> We’d accidentally wind up seeing %ebx as valid output of program string evaluation if we ever saw a program string that had “$ebx $ebx =” in it, but if we’re saying that %ebx isn’t really ever relevant to the walk anyway, then perhaps it doesn’t matter.

Patch-set: 4
Label: Code-Review=0