Update patch set 1
Patch Set 1:

i think this needs discussion on the mailing list.  we already have a depot_tools method for checking out all the repos to the right locations, and we have a `repo init && repo sync` method.  we don't want to break either of those.

also, if we're going to do submodules, it should really be for all the dependent repos and not just lss.

Patch-set: 1
Label: Code-Review=0