Update patch set 3
Patch Set 3:

I'm not sure if this is strictly *harmful*, but it definitely makes things confusing. I don't like the fact that it's stripping off the pathname in the `debug_file` method, that means it's losing data from the minidump.

Why not instead change `SimpleSymbolSupplier` to be able to use the module's `code_file` for lookup if `debug_file` is empty?
https://chromium.googlesource.com/breakpad/breakpad/+/8819ab0844819aa64a670efda85f5cddf40cf186/src/processor/simple_symbol_supplier.cc#156

Patch-set: 3