Update patch set 7
Patch Set 7: Code-Review+1

(5 comments)

rsesek@, can you check if this is acceptable for you?
we had some discussion locally today. things are quite complicated on Android because of all the relro packing craziness *, which has been the cause of all the back and forth on this CL between me and Toby.
There are various ways this CL could have been structured, but given the situation and the fact that I'm counting the hours before crashpad takes over, my suggestion is to go for the way which is less invasive and avoids touching any other existing part of breakpad.

WDYT?

* Last time I found out I knew this would have come back and wrote some detailed doc about that, see:
https://docs.google.com/presentation/d/1zhdACWvOwRMx3bJ4-3zRy2RzmFNaf-kq8eUVqEIRMVI/edit?usp=sharing

Patch-set: 7
Label: Code-Review=+1