Do not generate a microdump if there are no webview pointers on the stack.
The stack interest range is chosen by examining the process mappings to
find the one that contains a pointer that is known to be in the right
shared object (i.e. an appropriately chosen function pointer) passed from
the client.

If the extracted stack does not contain a pointer in this range, then we
assume that this is not a WebView crash, and do not generate a microdump.
If the stack extraction fails, we still generate a microdump (without a
stack).

BUG=

Change-Id: If19406a13168264f7751245fc39591bd6cdbf5df