Fix minidump name generation on Linux
This patch changes minidump filename generation on Linux so that the
name is in the same 8-4-4-4-12.dmp format as on the other platforms
and ensures that two crashes taken within the same second have different
minidump names. To fix the latter issue the random characters used in
the minidump filename are now read from /dev/urandom were possible or
generated via arc4random(). If neither is available we fall back to
regular rand() but mixing the address of an object to the current time
when generating the random seed to make it slightly less predictable.

BUG=727

Change-Id: I4d0d6b6444de246b534003242ed50cb5bc780984