Handle ntdll only emitting PUBLIC at func entry
This handles a case encountered in ntdll.dll symbols for Windows 7,
where a PUBLIC would be emitted for the entry point to the function. The
body of the function, however, is split in a PGO-ish fashion to another
remote location in the binary. Because of this, there were large gaps in
the RVA space that would be attributed to the "last" function that
happened to have an entry point before there. In practice, something
like this:

0x100 Func1
0x110 Func2
0x120 Func3
0x130 Func4
...
0x800 LaterFuncs

The bodies of Func1/2/3 can be implemented as a fastpath check, followed
by a jmp to somewhere in the range between 0x130 and 0x800. Because no
symbols were emitted for this range, everything is attributed to Func4,
causing crash misattribution.

In this CL, the change is: after emitting the entry point symbol, also
walk in the original OMAP entries through the untranslated binary, and
for each block until we resolve to a new symbol (via the same mechanism
as we found the entry point) emit another PUBLIC indicating that there's
another block that belongs to that symbol. This effective breaks up the
"0x130 - 0x800" range above.

R=mark@chromium.org
BUG=678874

Change-Id: Ib3741abab2e7158c81e3e34bca4340ce4d3153a1