Fix Windows CI build so dump_syms_unittest can find it's testdata
v2: Try to be smarter about how the path to testdata built into
dump_syms_unittest is determined

The dump_syms_unittest executable will be built in %CONFIGURATION%/,
relative to the .sln file, which is created by gyp in the same directory as
the .gyp file (assuming --generator-output isn't used).

dump_syms_unittest embeds a path for the testdata, relative to the directory
that contains the executable, currently '../testdata/'.

That relative path is only correct when dump_syms_unittest is built using
the .gyp file in the src/tools/windows/dump_syms directory (and
--generator-output isn't being used)

Instead build the absolute path to the testdata into dump_syms_unittest.

BUG=
Change-Id: I3886a4532f5fd5779ac7ad2045971d765000c9f9