Teach the ELF parser to handle multiple PT_NOTE phdrs.
It is legal for an ELF to contain multiple PT_NOTEs, and that is in
fact what lld's output looks like.

While here, simplify the code: the layout of Elf32_Nhdr and Elf64_Nhdr
is the same, so remove templating and code that extracts the elfclass
from the ELF file.

Testing: "make check" and breakpad_unittests when patched into
chromium.
Bug: 716484
Change-Id: I01d3f8679961e2cb7e789d4007de8914c6af357d