Raise file descriptor soft limit in signal handler
Processes may crash because they've hit their soft limit of file
descriptors. We've seen this in the wild, and the exhaustion of file
descriptors prevents crash information from being collected.

This change allows processes to set a soft limit below the hard limit so
the signal handler may raise it and report crash information.

BUG=chromium:583730
TEST=New unittest fails without this change, passes with.

Change-Id: Ic121e5f09b0c53eb73a89f98bc6a4e8c59549d88