Handle very large stack traces
The main motivation for this change is to handle very large stack traces,
normally the result of infinite recursion. This part is actually fairly simple,
relaxing a few self-imposed limits on how many frames we can unwind and
the max size for stack memory.

Relaxing these limits requires a stricter and more consistent checks for
stack unwinding. There are a number of unwinding invariants that apply
to all the platforms:

1. stack pointer (and frame pointer) must be within the stack memory range
   (frame pointer, if preset, must point to the right frame too)
2. unwinding must monotonically increase SP
   (except for the first frame unwind, this must be a strict increase)
3. Instruction pointer (return address) must point to a valid code location
4. stack pointer (and frame pointer) must be appropriately aligned

This change is focused on #2, which is enough to guarantee that the
unwinding doesn't get stuck in an infinite loop.

might be nice though). #4 is ABI specific and while it may be valuable in catching
suspicious is not relevant to the goal of this change.

more complex than just calling StackWalker::InstructionAddressSeemsValid() and
we don't want to drop parts of the callstack due to an overly conservative check.

Bug: 735989
Change-Id: I9aaba77c7fd028942d77c87d51b5e6f94e136ddd