Update patch set 1
Patch Set 1:

Is there a way to make the change to Makefile.in minimal? Updating aclocal.m4 and Makefile.in locally results in this change:

git diff --stat
 Makefile.in | 79 ++++++++++++++++++++++++++++++++++---------------------------------------------
 aclocal.m4  | 73 +++++++++++++++++++++++++++++++++++--------------------------------------

which seems excessive and not particularly reviewable. Is there a particular environment in which they should be regenerated?

Also, when you say "that's what the bots are for", is there a way to submit a tryjob from gerrit (I'm not seeing it...)? Or is there some other way to try a change before submitting it?

Patch-set: 1
Reviewer: Gerrit User 1133706 <1133706@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>