Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)

I'm happy to remove the implementation in favour of calling basename() and dirname(), but is it best to remove the unittests in that case? It seems we can't really test anything that might be implementation defined. I guess one feature of reimplementing is that there's no variation in the behaviour between platforms.

Patch-set: 3
Reviewer: Gerrit User 1133706 <1133706@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>