Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > this seems to have broken the unittests.  can you take a look ?
> > > https://travis-ci.org/google/breakpad/jobs/326887837
> > 
> > Having a look.
> 
> It looks like something on travis might have broken. I tried re-running the travis build with the last good version but got the same errors: https://travis-ci.org/lekv/breakpad/jobs/332398383
> 
> Locally all tests work, with and without this change. I will try to see if I can find out what's wrong with travis.

After some fiddling I was able to run the travis build locally in a Docker container and it succeeded as well. I'm no expert in Travis/Docker so I don't know what the best next steps are to debug this. However, I'm fairly confident that this is an issue on Travis's side and that this change did not cause the build to break.

Feel free to ping me if you think otherwise or if you have an idea what to check next. Should we start a thread on dev@ to get more eyes on this?

Patch-set: 2