Update patch set 2
Patch Set 2:

breakpad doesn't run on all those processors, so it's not a big deal if they aren't supported

then again, as i mentioned earlier, this is third_party code.  we copy this from upstream.  if you want to do a fresh import of the latest curl headers, feel free.

Patch-set: 2