Fix build of dump_syms as proposed by Benjamin Sergeant     (cf. https://goo.gl/a2z5xu )
Also includes a fix for macho_dump.cc so each SuperFatArch element is
converted one by one. Prevous code would convert only the first one
and then index the resulting pointer as it would be a pointer. More
seriously, the output of ConvertToFatArch would have been written to
an undefined location because object_files was not initialised, as
warned about by clang.

Change-Id: I0975f1049e1843b59f60e4bf2b3dbc769a35055c