Update patch set 1
Patch Set 1:

(5 comments)

I just had time to do a quick pass. One comment is that it seems like this does a lot of refactoring as well as making this change. Sometimes that is unavoidable but it is usually better to separate out the two, to make it easier to see what the actual changes are.

At the very least the CL description should mention that there is significant refactoring going on (as well as mentioning the rationale for this change).

Patch-set: 1
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>