Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (5 comments)
> > 
> > I just had time to do a quick pass. One comment is that it seems like this does a lot of refactoring as well as making this change. Sometimes that is unavoidable but it is usually better to separate out the two, to make it easier to see what the actual changes are.
> > 
> > At the very least the CL description should mention that there is significant refactoring going on (as well as mentioning the rationale for this change).
> 
> I get the feeling about moving code around, I'd definitely rather not have to do it for something that's conceptually so simple. I'm not sure if the ideal is to create two CLs out of these changes, but if others think so then I'm fine with doing so.
> 
> Depending on if I break it up, I'll put the rationale somewhere, as you suggest.
> 
> To briefly explain here, in case it's not self-evident why there is so much movement of code, most of what the PE-only symbol dumping is doing was already implemented in the PDBSourceLineWriter but that class had coupled everything in very tightly to the lifetime of some msdia resources. AFAICT the alternative to moving all the shared functionality out of PDBSourceLineWriter is making that class also function without a PDB, which felt to me like it would make the implementation too convoluted and the contract with consumers confusing.

For the code movement, one option is to do it first as a separate CL. But, sometimes that is a pain or just isn't worth it. But, if you do a CL which is refactoring and adding a feature I think the CL description *must* acknowledge that, preferably in enough detail to aid a reviewer or future reader in understanding the change.

But, I'll leave the rest of the reviewing to Mark and Ivan. I am highly supportive of the end-goal of this change.

Patch-set: 1
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>