Update patch set 4
Patch Set 4:

Hi, John. Thanks for the suggestion. I don’t think we want to carry a sample project like this. We have a hard enough time with bit-rot of the various build configurations as it is. This looks self-contained, why does it have to live in the Breakpad repository?

Patch-set: 4
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>