Update patch set 2
Patch Set 2:

Is this something that we want for all arm64 and not just Apple’s?

For that matter, the Apple doc says that this is restricted to arm64e. Can we be more surgical about our application too?

Patch-set: 2
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>