Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Is this something that we want for all arm64 and not just Apple’s?
> 
> For that matter, the Apple doc says that this is restricted to arm64e. Can we be more surgical about our application too?

I think arm64e is just Apple's name for it, but pointer authentication is specified in ARM64 v8.3-A and could eventually make it to other platforms.

That said, as far as I can tell, Apple is the only one producing chips that support it. It should be easy to limit this to iOS since we already have the OS+version strings in system_info (iOS 12+?). I don't think the cpu_subtype_t is stored in the minidump anywhere. We do record "arm64e" as the architecture in symbol files, though those are less convenient to access here.

Patch-set: 2
Reviewer: Gerrit User 1172779 <1172779@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>