Update patch set 2
Patch Set 2: Code-Review+1

OK. If we need to scope this to specific platforms, we can do that.

For the future: if there’s a way to access the appropriate mask on the client side, we should do that in Crashpad and store it somewhere useful in the dump. Then we could make this whole thing more deterministic.

Patch-set: 2
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1