Update patch set 2
Patch Set 2:

(1 comment)

The operative bit is fine, but by not doing a more global find-and-replace, you didn’t hit the comments that Xcode leaves behind. That, in turn, leaves the file in a non-canonical state such that anyone who opens it in the Xcode UI will have to suffer through Xcode “dirtying” the file to bring it up to snuff.

Patch-set: 2
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>