Remove dependency of uploader.mm on GTMLogger
The file GTMLogger shipped with breakpad is a copy of the version from
google_toolbox_for_mac. Having uploader.mm depend on GTMLogger causes
pain to iOS projects that want to integrate both google_toolbox_for_mac
and breakpad.

Since the file uploader.mm mixed uses of fprintf and GTMLogger to log
errors and warning, convert it to only use fprintf to stderr.

Bug: none
Change-Id: I68313ccf6951676a2859f44225281813722096ba