dump_syms: make breakpad sym files 10% - 30% smaller
Line table information in DWARF contains column info and the line table entries
will often have the same file and line, and differ only by the column. If the
address ranges of two lines are contiguous, we combine them. This ends up
saving 10% - 15% in the resulting size of the breakpad sym files with no loss
in accuracy.

This patch intercepts the function that grabs the line entries and appends
them so we don't end up with redundant line entries before we emit the
information to disk.

Patch by Greg Clayton

Change-Id: I694f5a60f10c162f360a6f6d25b323c454d1c985